+++ PassKeeper/app/src/main/res/layout/activity_agregar_password.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".GuardarPassword.Agregar_Password">
+    tools:context=".AgregarPassword">
 
     <View
         android:id="@+id/viewBlue"
@@ -49,7 +49,7 @@
 
         <com.google.android.material.card.MaterialCardView
             android:id="@+id/cardView2"
-            android:layout_width="328dp"
+            android:layout_width="332dp"
             android:layout_height="70dp"
             android:backgroundTint="@color/white"
             app:cardCornerRadius="30dp"
@@ -237,7 +237,7 @@
     </com.google.android.material.card.MaterialCardView>
 
     <com.google.android.material.button.MaterialButton
-        android:id="@+id/btnGuardarPassword"
+        android:id="@+id/btn_guardar"
         style="@style/Widget.MaterialComponents.Button"
         android:layout_width="231dp"
         android:layout_height="61dp"
